fixed more typos

+++ include/dsa.h
@@ -59,19 +59,19 @@ void *dsBSTSearchNode(dsBinaryTree_t *node, void *key,
 /***** Array Sorting Algorithms *****/
 
 //Insertion Sort
-void dsInsertionSort(void *arr, size_t size, size_t elem_size, int (*compare)(void *,void *));
+void dsInsertionSort(void *arr, int size, size_t elem_size, int (*compare)(void *,void *));
 
 //Selection Sort
-void dsSelectionSort(void *arr, size_t size, size_t elem_size, int (*compare)( void *,  void *));
+void dsSelectionSort(void *arr, int size, size_t elem_size, int (*compare)( void *,  void *));
 
 //Shell Sort
-void dsShellSort(void *arr, size_t size, size_t elem_size, int (*compare)(void *,void *));
+void dsShellSort(void *arr, int size, size_t elem_size, int (*compare)(void *,void *));
 
 //Quick Sort
-void dsQuickSort(void *arr, size_t size, size_t elem_size, int (*compare)(void *,void *));
+void dsQuickSort(void *arr, int size, size_t elem_size, int (*compare)(void *,void *));
 
 //Merge Sort
-void dsMergeSort(void *arr, size_t size, size_t elem_size, int (*compare)(void *, void *));
+void dsMergeSort(void *arr, int size, size_t elem_size, int (*compare)(void *, void *));
 
 
 #endif // !DSA_H_
